# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Swagger UI environment generated for the TM Forum Intent Management\
    \ API specification"
  version: "4.0.0"
  title: "Intent"
host: "serverRoot:8080"
basePath: "/tmf-api/intent/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /listener/intentCreateEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentCreateEvent"
      description: "Example of a client listener for receiving the notification IntentCreateEvent"
      operationId: "listenToIntentCreateEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentCreateEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/intentChangeEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentChangeEvent"
      description: "Example of a client listener for receiving the notification IntentChangeEvent"
      operationId: "listenToIntentChangeEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentChangeEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/intentDeleteEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentDeleteEvent"
      description: "Example of a client listener for receiving the notification IntentDeleteEvent"
      operationId: "listenToIntentDeleteEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentDeleteEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/intentReportCreateEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentReportCreateEvent"
      description: "Example of a client listener for receiving the notification IntentReportCreateEvent"
      operationId: "listenToIntentReportCreateEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentReportCreateEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/intentReportChangeEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentReportChangeEvent"
      description: "Example of a client listener for receiving the notification IntentReportChangeEvent"
      operationId: "listenToIntentReportChangeEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentReportChangeEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/intentReportDeleteEvent:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity IntentReportDeleteEvent"
      description: "Example of a client listener for receiving the notification IntentReportDeleteEvent"
      operationId: "listenToIntentReportDeleteEvent"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/IntentReportDeleteEvent"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  IntentCreateEvent:
    type: "object"
    properties:
      event:
        description: "The event payload linked to the involved resource object"
        $ref: "#/definitions/IntentCreateEventPayload"
      eventId:
        type: "string"
        description: "The identifier of the notification."
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurrence."
      eventType:
        type: "string"
        description: "The type of the notification."
      correlationId:
        type: "string"
        description: "The correlation id for this event."
      domain:
        type: "string"
        description: "The domain of the event."
      title:
        type: "string"
        description: "The title of the event."
      description:
        type: "string"
        description: "An explnatory of the event."
      priority:
        type: "string"
        description: "A priority."
      timeOcurred:
        type: "string"
        format: "date-time"
        description: "The time the event occured."
    description: "The notification data structure"
    example:
      eventId: "eventId"
      domain: "domain"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      correlationId: "correlationId"
      eventType: "eventType"
      event:
        intent:
          expression:
            iri: "iri"
            '@baseType': "@baseType"
            '@type': "@type"
            expressionLanguage: "expressionLanguage"
            '@schemaLocation': "http://example.com/aeiou"
          lifecycleStatus: "lifecycleStatus"
          validFor:
            startDateTime: "1985-04-12T23:20:50.52Z"
            endDateTime: "1985-04-12T23:20:50.52Z"
          '@type': "Intent"
          description: "description"
          creationDate: "2000-01-23T04:56:07.000+00:00"
          version: "version"
          statusChangeDate: "2000-01-23T04:56:07.000+00:00"
          '@baseType': "@baseType"
          lastUpdate: "2000-01-23T04:56:07.000+00:00"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
      title: "title"
      priority: "priority"
      timeOcurred: "2000-01-23T04:56:07.000+00:00"
  IntentCreateEventPayload:
    type: "object"
    properties:
      intent:
        description: "The involved resource data for the event"
        $ref: "#/definitions/Intent"
    description: "The event data structure"
    example:
      intent:
        expression:
          iri: "iri"
          '@baseType': "@baseType"
          '@type': "@type"
          expressionLanguage: "expressionLanguage"
          '@schemaLocation': "http://example.com/aeiou"
        lifecycleStatus: "lifecycleStatus"
        validFor:
          startDateTime: "1985-04-12T23:20:50.52Z"
          endDateTime: "1985-04-12T23:20:50.52Z"
        '@type': "Intent"
        description: "description"
        creationDate: "2000-01-23T04:56:07.000+00:00"
        version: "version"
        statusChangeDate: "2000-01-23T04:56:07.000+00:00"
        '@baseType': "@baseType"
        lastUpdate: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
  IntentChangeEvent:
    type: "object"
    properties:
      event:
        description: "The event payload linked to the involved resource object"
        $ref: "#/definitions/IntentChangeEventPayload"
      eventId:
        type: "string"
        description: "The identifier of the notification."
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurrence."
      eventType:
        type: "string"
        description: "The type of the notification."
      correlationId:
        type: "string"
        description: "The correlation id for this event."
      domain:
        type: "string"
        description: "The domain of the event."
      title:
        type: "string"
        description: "The title of the event."
      description:
        type: "string"
        description: "An explnatory of the event."
      priority:
        type: "string"
        description: "A priority."
      timeOcurred:
        type: "string"
        format: "date-time"
        description: "The time the event occured."
    description: "The notification data structure"
    example:
      eventId: "eventId"
      domain: "domain"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      correlationId: "correlationId"
      eventType: "eventType"
      event:
        intent:
          expression:
            iri: "iri"
            '@baseType': "@baseType"
            '@type': "@type"
            expressionLanguage: "expressionLanguage"
            '@schemaLocation': "http://example.com/aeiou"
          lifecycleStatus: "lifecycleStatus"
          validFor:
            startDateTime: "1985-04-12T23:20:50.52Z"
            endDateTime: "1985-04-12T23:20:50.52Z"
          '@type': "Intent"
          description: "description"
          creationDate: "2000-01-23T04:56:07.000+00:00"
          version: "version"
          statusChangeDate: "2000-01-23T04:56:07.000+00:00"
          '@baseType': "@baseType"
          lastUpdate: "2000-01-23T04:56:07.000+00:00"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
      title: "title"
      priority: "priority"
      timeOcurred: "2000-01-23T04:56:07.000+00:00"
  IntentChangeEventPayload:
    type: "object"
    properties:
      intent:
        description: "The involved resource data for the event"
        $ref: "#/definitions/Intent"
    description: "The event data structure"
    example:
      intent:
        expression:
          iri: "iri"
          '@baseType': "@baseType"
          '@type': "@type"
          expressionLanguage: "expressionLanguage"
          '@schemaLocation': "http://example.com/aeiou"
        lifecycleStatus: "lifecycleStatus"
        validFor:
          startDateTime: "1985-04-12T23:20:50.52Z"
          endDateTime: "1985-04-12T23:20:50.52Z"
        '@type': "Intent"
        description: "description"
        creationDate: "2000-01-23T04:56:07.000+00:00"
        version: "version"
        statusChangeDate: "2000-01-23T04:56:07.000+00:00"
        '@baseType': "@baseType"
        lastUpdate: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
  IntentDeleteEvent:
    type: "object"
    properties:
      event:
        description: "The event payload linked to the involved resource object"
        $ref: "#/definitions/IntentDeleteEventPayload"
      eventId:
        type: "string"
        description: "The identifier of the notification."
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurrence."
      eventType:
        type: "string"
        description: "The type of the notification."
      correlationId:
        type: "string"
        description: "The correlation id for this event."
      domain:
        type: "string"
        description: "The domain of the event."
      title:
        type: "string"
        description: "The title of the event."
      description:
        type: "string"
        description: "An explnatory of the event."
      priority:
        type: "string"
        description: "A priority."
      timeOcurred:
        type: "string"
        format: "date-time"
        description: "The time the event occured."
    description: "The notification data structure"
    example:
      eventId: "eventId"
      domain: "domain"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      correlationId: "correlationId"
      eventType: "eventType"
      event:
        intent:
          expression:
            iri: "iri"
            '@baseType': "@baseType"
            '@type': "@type"
            expressionLanguage: "expressionLanguage"
            '@schemaLocation': "http://example.com/aeiou"
          lifecycleStatus: "lifecycleStatus"
          validFor:
            startDateTime: "1985-04-12T23:20:50.52Z"
            endDateTime: "1985-04-12T23:20:50.52Z"
          '@type': "Intent"
          description: "description"
          creationDate: "2000-01-23T04:56:07.000+00:00"
          version: "version"
          statusChangeDate: "2000-01-23T04:56:07.000+00:00"
          '@baseType': "@baseType"
          lastUpdate: "2000-01-23T04:56:07.000+00:00"
          name: "name"
          id: "id"
          href: "http://example.com/aeiou"
          '@schemaLocation': "http://example.com/aeiou"
      title: "title"
      priority: "priority"
      timeOcurred: "2000-01-23T04:56:07.000+00:00"
...
  IntentReportCreateEvent:
    type: "object"
    properties:
      event:
        description: "The event payload linked to the involved resource object"
        $ref: "#/definitions/IntentReportCreateEventPayload"
      eventId:
        type: "string"
        description: "The identifier of the notification."
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurrence."
      eventType:
        type: "string"
        description: "The type of the notification."
      correlationId:
        type: "string"
        description: "The correlation id for this event."
      domain:
        type: "string"
        description: "The domain of the event."
      title:
        type: "string"
        description: "The title of the event."
      description:
        type: "string"
        description: "An explnatory of the event."
      priority:
        type: "string"
        description: "A priority."
      timeOcurred:
        type: "string"
        format: "date-time"
        description: "The time the event occured."
    description: "The notification data structure"
    example:
      eventId: "eventId"
      domain: "domain"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      correlationId: "correlationId"
      eventType: "eventType"
      event:
        intentReport:
          expression:
            iri: "iri"
            '@baseType': "@baseType"
            '@type': "@type"
            expressionLanguage: "expressionLanguage"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          validFor:
            startDateTime: "1985-04-12T23:20:50.52Z"
            endDateTime: "1985-04-12T23:20:50.52Z"
          '@type': "@type"
          name: "name"
          description: "description"
          id: "id"
          href: "http://example.com/aeiou"
          creationDate: "2000-01-23T04:56:07.000+00:00"
          '@schemaLocation': "http://example.com/aeiou"
          intent:
            expression:
              iri: "iri"
              '@baseType': "@baseType"
              '@type': "@type"
              expressionLanguage: "expressionLanguage"
              '@schemaLocation': "http://example.com/aeiou"
            lifecycleStatus: "lifecycleStatus"
            validFor:
              startDateTime: "1985-04-12T23:20:50.52Z"
              endDateTime: "1985-04-12T23:20:50.52Z"
            '@type': "Intent"
            description: "description"
            creationDate: "2000-01-23T04:56:07.000+00:00"
            version: "version"
            statusChangeDate: "2000-01-23T04:56:07.000+00:00"
            '@baseType': "@baseType"
            lastUpdate: "2000-01-23T04:56:07.000+00:00"
            name: "name"
            id: "id"
            href: "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
      title: "title"
      priority: "priority"
      timeOcurred: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "string"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "string"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"